refactor: split runtime persistence helpers
Extract read-side run serialization and step-execution input/result helpers out of the large runtime modules. Keep the public runtime behavior unchanged while making the command flow easier to follow for new contributors.
